move details into own component
