+++ frontend/src/components/index.ts
@@ -92,6 +92,9 @@ import("./crawl-list").then(({ CrawlListItem, CrawlList }) => {
 import("./section-heading").then(({ SectionHeading }) => {
   customElements.define("btrix-section-heading", SectionHeading);
 });
+import("./config-details").then(({ ConfigDetails }) => {
+  customElements.define("btrix-config-details", ConfigDetails);
+});
 
 customElements.define("btrix-alert", Alert);
 customElements.define("btrix-input", Input);
